+++ src/table/hooks/useEditableRow.ts
@@ -1,0 +1,79 @@
+import { useState, useMemo } from 'react';
+import get from 'lodash/get';
+import { PrimaryTableProps } from '../interface';
+import { validate } from '../../form/formModel';
+import { PrimaryTableRowEditContext, TableRowData } from '../type';
+import { ErrorListType } from '../PrimaryTable';
+import { AllValidateResult } from '../../form';
+import { getEditableKeysMap } from '../utils';
+
+const cellRuleMap = new Map<any, PrimaryTableRowEditContext<TableRowData>[]>();
+
+export function useEditableRow(props: PrimaryTableProps) {
+  const { editableRowKeys } = props;
+  const [errorListMap, setErrorListMap] = useState<ErrorListType>({});
+  const editableKeysMap = useMemo(
+    () => editableRowKeys && getEditableKeysMap(editableRowKeys, props.data, props.rowKey || 'id'),
+    [editableRowKeys, props.data, props.rowKey],
+  );
+
+  const validateRowData = (rowValue: any) => {
+    const rowRules = cellRuleMap.get(rowValue);
+    const list = rowRules.map(
+      (item) =>
+        new Promise<PrimaryTableRowEditContext<TableRowData> & { errorList: AllValidateResult[] }>((resolve) => {
+          const { value, col } = item;
+          if (!col.edit || !col.edit.rules || !col.edit.rules.length) {
+            resolve({ ...item, errorList: [] });
+            return;
+          }
+          validate(value, col.edit.rules).then((r) => {
+            resolve({ ...item, errorList: r.filter((t) => !t.result) });
+          });
+        }),
+    );
+    Promise.all(list).then((results) => {
+      const errors = results.filter((t) => t.errorList.length);
+      const errorMap: ErrorListType = {};
+      errors.forEach(({ row, col, errorList }) => {
+        const rowValue = get(row, props.rowKey || 'id');
+        const key = [rowValue, col.colKey].join();
+        errorMap[key] = errorList;
+      });
+      setErrorListMap(errorMap);
+      // 缺少校验文本显示
+      props.onRowValidate?.({ trigger: 'parent', result: errors });
+    });
+  };
+
+  const onRuleChange = (context: PrimaryTableRowEditContext<TableRowData>) => {
+    // 编辑行，预存校验信息，方便最终校验
+    if (props.editableRowKeys) {
+      const rowValue = get(context.row, props.rowKey || 'id');
+      const rules = cellRuleMap.get(rowValue);
+      if (rules) {
+        const index = rules.findIndex((t) => t.col.colKey === context.col.colKey);
+        if (index === -1) {
+          rules.concat(context);
+        } else {
+          rules[index].value = context.value;
+        }
+        cellRuleMap.set(rowValue, rules);
+      } else {
+        cellRuleMap.set(rowValue, [context]);
+      }
+    }
+  };
+
+  const clearValidateData = () => {
+    setErrorListMap({});
+  };
+
+  return {
+    errorListMap,
+    editableKeysMap,
+    validateRowData,
+    clearValidateData,
+    onRuleChange,
+  };
+}
